+++ archon-ui-main/src/features/projects/documents/services/index.ts
@@ -2,10 +2,7 @@
  * Document Services
  *
  * Service layer for document operations.
- * Currently using projectService from the main services directory.
- * This file exists to maintain structural consistency.
+ * Part of the vertical slice architecture migration.
  */
 
-// Document services are currently handled by projectService
-// This file is a placeholder for future document-specific services
-export {};
+export { documentService } from './documentService';
