+++ archon-ui-main/src/features/projects/tasks/services/index.ts
@@ -1,0 +1,8 @@
+/**
+ * Task Services
+ *
+ * Service layer for task operations.
+ * Part of the vertical slice architecture migration.
+ */
+
+export { taskService } from './taskService';
